WEB: Register Device fix remove button, confirm removal of device
git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@318 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18

+++ web/webapp/js/registerARD.js
@@ -16,14 +16,16 @@ $(function() {
 	});
 	
 	$("#device-remove").click(function() {
-		$.ajax({
-			type: "DELETE",
-			url: ctx+"/api/ard",
-			contentType : "application/json; charset=utf-8",
-			complete: ajaxMessageClosure(messagable, function(removed, message) {
-				showRemoveButton(!removed);
-			})
-		});
+		if (confirm("Are you sure you want to unlink this device from your account?")) {
+			$.ajax({
+				type: "DELETE",
+				url: ctx+"/api/ard",
+				contentType : "application/json; charset=utf-8",
+				complete: ajaxMessageClosure(messagable, function(removed, message) {
+					showRemoveButton(!removed);
+				})
+			});
+		}
 	});
 	
 	function updateARD() {
